refactor: reorganize directories
