+++ nonproject/header/domain/commands/operator/Kick.hpp
